Rename module operation_meta
