+++ aquadoggo/src/graphql/client/static_types/operation_meta.rs
